refactor: dto 변환 메서드 위치 변경
- 별도의 객체가 매핑하던 방식에서 dto 스스로 변환할 수 있도록 합니다.
