+++ src/main/java/roomescape/presentation/dto/ReservationResponseDto.java
@@ -1,6 +1,7 @@
 package roomescape.presentation.dto;
 
 import java.time.LocalDate;
+import roomescape.business.domain.reservation.Reservation;
 
 public record ReservationResponseDto(
         long id,
@@ -9,4 +10,14 @@ public record ReservationResponseDto(
         ReservationTimeResponseDto time,
         ReservationThemeResponseDto theme
 ) {
+
+    public static ReservationResponseDto toResponse(Reservation reservation) {
+        return new ReservationResponseDto(
+                reservation.getId(),
+                reservation.getName(),
+                reservation.getDate(),
+                ReservationTimeResponseDto.toResponse(reservation.getTime()),
+                ReservationThemeResponseDto.toResponse(reservation.getTheme())
+        );
+    }
 }
