+++ src/main/java/roomescape/presentation/dto/ReservationThemeResponseDto.java
@@ -1,9 +1,20 @@
 package roomescape.presentation.dto;
 
+import roomescape.business.domain.reservation.ReservationTheme;
+
 public record ReservationThemeResponseDto(
         long id,
         String name,
         String description,
         String thumbnail
 ) {
+
+    public static ReservationThemeResponseDto toResponse(ReservationTheme reservationTheme) {
+        return new ReservationThemeResponseDto(
+                reservationTheme.getId(),
+                reservationTheme.getName(),
+                reservationTheme.getDescription(),
+                reservationTheme.getThumbnail()
+        );
+    }
 }
