--- src/main/java/roomescape/presentation/mapper/ReservationMapper.java
@@ -1,20 +1,0 @@
-package roomescape.presentation.mapper;
-
-import roomescape.business.domain.reservation.Reservation;
-import roomescape.presentation.dto.ReservationResponseDto;
-
-public final class ReservationMapper {
-
-    private ReservationMapper() {
-    }
-
-    public static ReservationResponseDto toResponse(Reservation reservation) {
-        return new ReservationResponseDto(
-                reservation.getId(),
-                reservation.getName(),
-                reservation.getDate(),
-                ReservationTimeMapper.toResponse(reservation.getTime()),
-                ReservationThemeMapper.toResponse(reservation.getTheme())
-        );
-    }
-}
