--- src/main/java/roomescape/presentation/mapper/ReservationTimeMapper.java
@@ -1,17 +1,0 @@
-package roomescape.presentation.mapper;
-
-import roomescape.business.domain.reservation.ReservationTime;
-import roomescape.presentation.dto.ReservationTimeResponseDto;
-
-public final class ReservationTimeMapper {
-
-    private ReservationTimeMapper() {
-    }
-
-    public static ReservationTimeResponseDto toResponse(ReservationTime reservationTime) {
-        return new ReservationTimeResponseDto(
-                reservationTime.getId(),
-                reservationTime.getStartAt()
-        );
-    }
-}
